Added the third impl

+++ ch03/3.4/weatherenrich/src/main/java/weatherenrich/events/EnrichedEvent.java
@@ -5,7 +5,7 @@ public class EnrichedEvent extends RawEvent {
   public final EnrichedContext context;
 
   public EnrichedEvent(RawEvent rawEvent,
-    Float temperature, String[] conditions) {
+    Float temperature, String[] conditions) {                          // a
 
     this.subject = rawEvent.subject;
     this.verb = rawEvent.verb;
@@ -14,7 +14,7 @@ public class EnrichedEvent extends RawEvent {
       rawEvent.context.timestamp, temperature, conditions);
   }
 
-	public static final class EnrichedContext extends Context {
+  public static final class EnrichedContext extends Context {          // b
     public final String timestamp;
     public final Weather weather;
 
@@ -26,8 +26,8 @@ public class EnrichedEvent extends RawEvent {
     }
 
     public static final class Weather {
-      public final Float temperature; // TODO: check precision
-      public final String[] conditions; // TODO: Array or List?
+      public final double temperature;
+      public final String[] conditions;
 
       public Weather(Float temperature, String[] conditions) {
         this.temperature = temperature;
